Update RECOM_UBCF.R
https://github.com/mhahsler/recommenderlab/issues/43#issue-734093255

.knn <- function(sim, k){...}
When any user for prediction has fewer neighbors than k, the knn is a list instead of a matrix. The added if block checks whether knn is a list and if it is a list, stop the code and return a message indicating the first user whose number of neighbors is less than k.