report the user having the fewest # of neighbors
Note: when any user for prediction has fewer neighbors than k, the knn is a list instead of a matrix. The added if block below checks whether knn is a list and if it is a list, stop the code and return a message indicating the user who has the fewest neighbors. The first is returned when there are multiple such users.